debug create filter

+++ scripts/ASBcalling/BAD_annotation.py
@@ -1,7 +1,7 @@
 import json
+import os
 import sys
-from scripts.HELPERS.paths import create_badmaps_path_function, get_ending
-from scripts.HELPERS.paths_for_components import badmaps_dict_path
+from scripts.HELPERS.paths import create_badmaps_path_function, get_ending, get_merged_badmaps_dict_path
 from scripts.HELPERS.helpers import callers_names, unpack, pack, Intersection, UnpackBadSegments, segmentation_states
 
 
@@ -14,18 +14,20 @@ def make_reverse_dict(dictionary):
     return new_dict
 
 
-def main(key):
+def main(key, remade=True):
     table_annotated = key + get_ending("annotation")
     output = key + get_ending("BAD")
 
-    with open(badmaps_dict_path, "r") as read_file:
-        d = json.loads(read_file.readline())
+    with open(get_merged_badmaps_dict_path(remade=remade), "r") as read_file:
+        d = json.load(read_file)
         rev_d = make_reverse_dict(d)
 
     badmap_file_name = rev_d[key]
 
     print('Now doing {} \n with BAD map file {}'.format(table_annotated, badmap_file_name))
     badmap_file_path = create_badmaps_path_function(badmap_file_name, valid=True)
+    if not os.path.isfile(badmap_file_path):
+        exit(0)
     with open(badmap_file_path, 'r') as badmap_file, open(output, 'w') as out, open(table_annotated, 'r') as table_file:
         out.write(pack(['#chr', 'pos', 'ID', 'ref', 'alt', 'ref_read_counts', 'alt_read_counts',
                         'repeat_type'] + callers_names + ['BAD'] + ["Q{:.2f}".format(x) for x in segmentation_states] +
